Split interpreter.ts into smaller files
